We're now supporting passing single values to the fluid type calculator, and returning clamp values based on a set of defaults: minFontSizeFactor, maxFontSizeFactor and scaleFactor.
So that themes can opt-in to the system, typography.fluid should be true, in which case default values will be used, or an object of viewport width, minFontSizeFactor, maxFontSizeFactor and scaleFactor.

The values in fontSizes will take precedence, then, if they're not there, we'll calculate a min and max font size based on the scale value.

The calculations for min() and max() have been removed since we'll always have a min and max font size value. This means we're returning a clamp() value always. I think that's okay as it also removes any unexpected side effects.